requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
langchain==0.1.7
langchain-openai==0.0.6
python-dotenv==1.0.1
cors==1.0.1
